+++ src/amg/azext_amg/commands.py
@@ -58,7 +58,7 @@ def load_command_table(self, _):
         g.custom_command('list', 'list_service_account_tokens')
         g.custom_command('delete', 'delete_service_account_token')
 
-    with self.command_group('grafana integrations monitor') as g:
-        g.custom_command('add', 'link_monitor')
-        g.custom_command('list', 'list_monitors')
-        g.custom_command('delete', 'unlink_monitor')
+    with self.command_group('grafana integration monitor') as g:
+        g.custom_command('add', 'link_monitor', is_preview=True)
+        g.custom_command('list', 'list_monitors', is_preview=True)
+        g.custom_command('delete', 'unlink_monitor', is_preview=True)